requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-crispy-forms==2.0
crispy-bootstrap5==0.7
python-dotenv==1.0.0
django-extensions==3.2.3